requirements.txt
asgiref==3.2.10
boto3==1.14.33
botocore==1.17.33
Django==3.0.9
django-cleanup==5.0.0
django-filter==2.3.0
djangorestframework==3.10.0
django-heroku==0.3.1
django-cors-headers==3.4.0
django-storages==1.9.1
gunicorn
Markdown==3.2.2
Pillow==7.2.0
python-dotenv==0.14.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
